Code cleanup (#338)
Re-organize some code. Recently worked in this folder and realized it
was confusing to navigate a lot of where what was.
