+++ .changeset/beige-toys-camp.md
@@ -1,0 +1,5 @@
+---
+"@zus-health/ctw-component-library": patch
+---
+
+Code cleanup.
