+++ src/components/content/forms/schemas/condition-schema.tsx
@@ -1,6 +1,6 @@
-import type { FormEntry } from "../../core/form/drawer-form-with-fields";
+import type { FormEntry } from "../../../core/form/drawer-form-with-fields";
 import Zod, { RefinementCtx, z } from "zod";
-import { ConditionsAutoComplete } from "./conditions-autocomplete";
+import { ConditionsAutoComplete } from "../conditions-autocomplete";
 import { ConditionModel } from "@/fhir/models/condition";
 
 export const getAddConditionData = ({
